Implement data models in dbt

+++ scraper/make_sqlite.py
@@ -1,26 +1,17 @@
 """Use the saved review data to build an analytics-ready SQLite db."""
 import argparse
-import datetime
 import gzip
 import json
 import multiprocessing
 import random
 import sqlite3
 from pathlib import Path
-from typing import Any, Generator, Iterable, Union
+from typing import Any, Generator, Iterable
 
-from bs4 import BeautifulSoup
-from pydantic import BaseModel, root_validator, validator
+from .models import Review
 from tqdm import tqdm
 
-from ._utils import FIRST_BEST_NEW_MUSIC, REVIEWS_SAVE_PATH, SQL_FILES, SQLITE_SAVE_PATH
-
-
-def unique(l: list[Any]) -> list[Any]:
-    """Get unique items and retain order. Thx stackoverflow."""
-    seen = set()
-    seen_add = seen.add
-    return [x for x in l if not (x in seen or seen_add(x))]
+from ._utils import dbt, FIRST_BEST_NEW_MUSIC, REVIEWS_SAVE_PATH, SQLITE_SAVE_PATH
 
 
 def chunker(seq: Iterable, size: int) -> Generator:
@@ -31,6 +22,11 @@ def chunker(seq: Iterable, size: int) -> Generator:
 def parse_args() -> argparse.Namespace:
     """Make the parser for the command line arguments."""
     parser = argparse.ArgumentParser(description=__doc__)
+    parser.add_argument(
+        "--no-dbt",
+        action="store_true",
+        help="Option to Skip DBT steps (assuming theyre run separately).",
+    )
     parser.add_argument(
         "--in",
         dest="in_",
@@ -48,11 +44,6 @@ def parse_args() -> argparse.Namespace:
         "--single", type=Path, help="Run only a single file. Useful for debugging."
     )
     parser.add_argument(
-        "--sample",
-        type=int,
-        help="Run a random sample of files. Also useflul for debugging.",
-    )
-    parser.add_argument(
         "--procs",
         type=lambda x: multiprocessing.cpu_count() if x == "max" else int(x),
         default=1,
@@ -60,203 +51,6 @@ def parse_args() -> argparse.Namespace:
     )
     args = parser.parse_args()
     return args
-
-
-class Artist(BaseModel):
-    name: str
-    url: Union[str, None]
-
-    @validator("url", always=True)
-    def check_url_startswith_artist(cls, v):
-        if v is not None and not v.startswith("/artists/"):
-            raise ValueError(f"{v} is not a valid URL.")
-        return v
-
-    @validator("name", always=True)
-    def check_name_has_value(cls, v):
-        assert v
-        return v
-
-    @property
-    def artist_id(self) -> str:
-        if self.url is None:
-            return "various-" + self.name.replace(" ", "-").lower()
-        return self.url.lstrip("/artists/")
-
-    def __hash__(self) -> int:
-        return (self.name, self.url).__hash__()
-
-
-class Tombstone(BaseModel):
-    title: str
-    release_years: Union[list[int], None]
-    score: float
-    bnm: bool
-
-    @validator("title", always=True)
-    def check_title_has_value(cls, v):
-        assert v
-        return v
-
-    @validator("score", always=True)
-    def check_score_bounds(cls, v):
-        assert v >= 0 and v <= 10
-        return v
-
-    @classmethod
-    def from_soup(cls, soup: BeautifulSoup) -> "Tombstone":
-        return cls(
-            title=cls.get_title(soup),
-            release_years=cls.get_release_years(soup),
-            score=cls.get_score(soup),
-            bnm=cls.get_bnm(soup),
-        )
-
-    @staticmethod
-    def get_title(soup: BeautifulSoup) -> str:
-        span = soup.find("h1", {"class": "single-album-tombstone__review-title"})
-        assert span is not None
-        return span.text.strip()
-
-    @staticmethod
-    def get_release_years(soup: BeautifulSoup) -> list[int]:
-        span = soup.find("span", {"class": "single-album-tombstone__meta-year"})
-        assert span is not None
-        span_text = span.text.replace("•", "").strip()
-
-        # some reviews do not publish a release date
-        if not span_text:
-            return None
-        return unique([int(i.split()[-1]) for i in span_text.split("/")])
-
-    @staticmethod
-    def get_score(soup: BeautifulSoup) -> list[float]:
-        span = soup.find("span", {"class": "score"})
-        assert span is not None
-        return float(span.text.strip())
-
-    @staticmethod
-    def get_bnm(soup: BeautifulSoup) -> list[bool]:
-        return soup.find("p", {"class": "bnm-txt"}) is not None
-
-
-class Review(BaseModel):
-    is_multi_review: bool
-    artists: list[Artist]
-    body: str
-    labels: list[str]
-    genres: list[str]
-    pub_date: datetime.datetime
-    authors: list[str]
-    tombstones: list[Tombstone]
-
-    @validator("pub_date", always=True)
-    def check_pub_date_value(cls, v):
-        assert v >= datetime.datetime(1999, 1, 1)
-        return v
-
-    @validator("artists", "authors", "tombstones", always=True)
-    def check_has_values(cls, v):
-        assert v
-        return v
-
-    @root_validator
-    def check_body_except_jet_shine_on(cls, values):
-        """Make an assertion about the review body.
-
-        Pitchfork hilariously reviewed Shine on by Jet with a video of a chimpanzee
-        peeing, so need to except that.
-        """
-        first_title = values["tombstones"][0].title
-        first_artist = values["artists"][0].name
-        if first_title == "Shine On" and first_artist == "Jet":
-            values["body"] = "https://www.youtube.com/watch?v=SvZmRv6U_s0&t=1s"
-            return values
-        assert values["body"]
-        return values
-
-    @classmethod
-    def from_html(cls, html: str) -> "Review":
-        """Create a Review object from an HTML string."""
-        soup = BeautifulSoup(html, "lxml")
-        return cls(
-            is_multi_review=cls.check_multi_review(soup),
-            artists=cls.get_artists(soup),
-            genres=cls.get_genres(soup),
-            body=cls.get_review_body(soup),
-            labels=cls.get_release_labels(soup),
-            pub_date=cls.get_pub_date(soup),
-            authors=cls.get_authors(soup),
-            tombstones=cls.get_tombstones(soup),
-        )
-
-    @staticmethod
-    def check_multi_review(soup: BeautifulSoup) -> bool:
-        """Return True if the review is for a multi-album release."""
-        return soup.find("div", {"class": "multi-tombstone-widget"}) is not None
-
-    @staticmethod
-    def get_artists(soup: BeautifulSoup) -> list[Artist]:
-        ul = soup.find("ul", {"class": ["artist-list"]})
-        assert ul is not None
-        return unique(
-            [
-                Artist(
-                    url=li.find("a")["href"] if li.find("a") else None,
-                    name=li.text.strip(),
-                )
-                for li in ul.findAll("li")
-            ]
-        )
-
-    @staticmethod
-    def get_pub_date(soup: BeautifulSoup) -> datetime.datetime:
-        time_ = soup.find("time", {"class": "pub-date"})
-        return datetime.datetime.fromisoformat(time_["datetime"])
-
-    @classmethod
-    def get_release_labels(cls, soup: BeautifulSoup) -> list[str]:
-        ul = soup.find("ul", {"class": ["labels-list"]})
-        return unique([i.text.strip() for i in ul.findAll("li")])
-
-    @staticmethod
-    def get_review_body(soup: BeautifulSoup) -> str:
-        div = soup.find("div", {"class": "review-detail__article-content"})
-        ps = []
-        for el in div.find_all(["p", "hr"]):
-            if el.name == "hr":
-                break
-            ps.append(el.text.strip())
-        return "\n\n".join(ps)
-
-    @staticmethod
-    def get_genres(soup: BeautifulSoup) -> list[str]:
-        ul = soup.find("ul", {"class": "genre-list"})
-        # some don't have genres
-        return unique([] if ul is None else [i.text.strip() for i in ul.findAll("li")])
-
-    @staticmethod
-    def get_authors(soup: BeautifulSoup) -> list[str]:
-        ul = soup.find("ul", {"class": "authors-detail"})
-        return unique(
-            [
-                i.text.strip()
-                for i in ul.findAll("a", {"class": "authors-detail__display-name"})
-            ]
-        )
-
-    @classmethod
-    def get_tombstones(cls, soup: BeautifulSoup) -> list[Tombstone]:
-        """Get the tombstone metadata, with multi album logic."""
-        if cls.check_multi_review(soup):
-            ul = soup.find("ul", {"class": "review-tombstones"})
-            return [
-                Tombstone.from_soup(div)
-                for div in ul.find_all("div", {"class": "single-album-tombstone"})
-            ]
-        else:
-            div = soup.find("div", {"class": "single-album-tombstone"})
-            return [Tombstone.from_soup(div)]
 
 
 def insert_many(
@@ -282,7 +76,6 @@ def insert_many(
 
     try:
         db.executemany(sql, vals)
-        # db.commit()
     except sqlite3.IntegrityError:
         if integrity_handler == "warn":
             print(f"Warning: Integrity error ignored on {table_name}.")
@@ -370,22 +163,25 @@ if __name__ == "__main__":
     else:
         assert args.in_.exists()
 
-    if args.out.exists():
+    if args.out.exists() and not args.no_dbt:
         args.out.unlink()
 
     review_jsons = (
         tuple(args.in_.glob("*.json.gz")) if not args.single else [args.single]
     )
 
-    if args.sample is not None:
-        review_jsons = random.sample(review_jsons, args.sample)
+    if not args.no_dbt:
+        print("Executing DBT clean...")
+        dbt("clean")
+        print()
+
+        # build tables
+        print("Executing DBT run...")
+        dbt("run")
+        print()
 
     # shared across later lines. idc about closing it, this is sqlite.
     db = sqlite3.connect(args.out, timeout=10000, check_same_thread=False)
-
-    # build tables
-    for fpath in SQL_FILES["ddl"]:
-        db.execute(fpath.read_text())
 
     def f(fpath: Path) -> tuple[str, Review]:
         with gzip.open(fpath, "rb") as f:
@@ -399,16 +195,19 @@ if __name__ == "__main__":
 
     chunks = list(chunker(review_jsons, min(1000, len(review_jsons))))
 
-    print(f"Multiprocessing on {len(chunks)} chunks of len={len(chunks[0])}")
+    print(f"Inserting data in {len(chunks)} chunks of len={len(chunks[0])}")
     with multiprocessing.Pool(args.procs) as pool:
         for chunk in tqdm(chunks):
             results = pool.map(f, chunk)
             for url, review in results:
                 insert_review(db, url, review)
 
-    # indices/views
-    for fpath in SQL_FILES["index"] + SQL_FILES["view"]:
-        db.execute(fpath.read_text())
-
     db.commit()
     db.close()
+
+    if not args.no_dbt:
+        print("\nExecuting DBT test...")
+        dbt("test")
+        print()
+
+    print("All good!")
